[IMP] account_loan: Be loan donor (allow negative amounts)
The usual case is to receive a loan from a bank institution, but when
you have a multi-company environment, one of the company may give a loan
to another, and for registering the loans on the donor part, you may
be able to enter negative amounts and configure the corresponding
accounts to receive money instead.

This commit allows to post the proper entries, checking the sign of the
amounts and acting in consequence.